feat: add rpc, relayer, sigServer cli cmd
