+++ offchain-modules/package.json
@@ -1,6 +1,6 @@
 {
   "name": "force-bridge",
-  "version": "0.0.1",
+  "version": "0.0.7",
   "description": "",
   "private": true,
   "workspaces": [
