+++ offchain-modules/packages/app-cli/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@force-bridge/cli",
   "main": "dist",
-  "version": "0.0.2",
+  "version": "0.0.10",
   "bin": {
     "forcecli": "./dist/index.js"
   },
@@ -13,6 +13,9 @@
     "test": "yarn run build"
   },
   "dependencies": {
-    "@force-bridge/x": "^0.0.2"
+    "@force-bridge/app-multisign-server": "*",
+    "@force-bridge/app-relayer": "*",
+    "@force-bridge/app-rpc-server": "*",
+    "@force-bridge/x": "*"
   }
 }
