+++ offchain-modules/packages/app-multisign-server/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@force-bridge/app-multisign-server",
   "main": "dist",
-  "private": true,
   "scripts": {
     "build": "yarn run clean && yarn run compile",
     "clean": "rimraf -rf ./dist",
@@ -10,7 +9,7 @@
     "test": "yarn run build"
   },
   "dependencies": {
-    "@force-bridge/x": "^0.0.2"
+    "@force-bridge/x": "^0.0.10"
   },
-  "version": "0.0.2"
+  "version": "0.0.10"
 }
