+++ offchain-modules/packages/x/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@force-bridge/x",
-  "version": "0.0.2",
+  "version": "0.0.10",
   "description": "",
   "main": "./dist/index.js",
   "scripts": {
